Moves Instances interface methods out of aws.go
